Another attempt at not deleting the output folder on Linux. Moved project to version 2.0.4.

+++ src/main/java/eu/isas/searchgui/SearchHandler.java
@@ -2316,7 +2316,9 @@ public class SearchHandler {
                     }
                 }
 
-                Util.deleteDir(outputTempFolder);
+                if (!outputFolder.getAbsolutePath().equals(outputTempFolder.getAbsolutePath())) {
+                    Util.deleteDir(outputTempFolder);
+                }
 
                 finished = true;
                 if (!waitingHandler.isRunCanceled()) {
